Update src/data_structure/bitset.rs
looks good :)

+++ src/data_structure/bitset.rs
@@ -54,7 +54,7 @@ pub mod bitset {
             let (data_index, bit_index) = get_bit_position(index);
             assert!(self.data.len() > data_index);
             if value {
-                self.data[data_index] |= 1 << (bit_index as u64);
+                self.data[data_index] |= 1u64 << bit_index;
             } else {
                 let tmp = MAXIMUM ^ 1 << (bit_index as u64);
                 self.data[data_index] &= tmp;
